FEAT: move grid spacing functions to new file
